requirements.txt
attrs==25.1.0
bokeh==2.4.3
branca==0.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
folium==0.19.5
fonttools==4.56.0
h11==0.14.0
idna==3.10
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
mysql-connector==2.2.9
numpy==1.23.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pyparsing==3.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
selenium==4.29.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
tornado==6.4.2
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
websocket-client==1.8.0
wsproto==1.2.0
xyzservices==2025.1.0
